Add files via upload
This project is about the design of a camera sensor to be integrated into a WSN capable of detection of unattended or missing objects in the area under it's surveillance. The results of such events is then published on a Thingspeak channel to enable remote monitoring. The instructions file clearly explains how to implement the project. The project has been implemented utilizing a Raspberry Pi 2 model B.